Add rough time series rendering for pyramids
