+++ components/dataset.js
@@ -29,6 +29,8 @@ const DATASETS = [
   'gs://leap-persistent-ro/data-library-manual/CESM.zarr',
   // missing CF axes
   'https://cpdataeuwest.blob.core.windows.net/cp-cmip/version1/data/MACA/CMIP.NCC.NorESM2-LM.historical.r1i1p1f1.day.MACA.tasmax.zarr',
+  // 4D pyramid
+  'https://storage.googleapis.com/carbonplan-maps/v2/demo/4d/tavg-prec-month/',
 ]
 
 const sx = {
